Update structure for project
