--- seagrass_api/main.py
@@ -1,20 +1,0 @@
-from fastapi import FastAPI
-from seagrass_api.api import api_router
-from fastapi.middleware.cors import CORSMiddleware
-
-app = FastAPI()
-
-app.include_router(api_router)
-
-# Allowing all middleware is optional, but good practice for dev purposes
-app.add_middleware(
-    CORSMiddleware,
-    allow_origins=["*"],  # Allows all origins
-    allow_credentials=True,
-    allow_methods=["*"],  # Allows all methods
-    allow_headers=["*"],  # Allows all headers
-)
-
-if __name__ == "__main__":
-    import uvicorn
-    uvicorn.run("taxifare.main:app", host="0.0.0.0", port=8000, reload=True)
